% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxNMF-package.R
\docType{package}
\name{coxNMF-package}
\alias{coxNMF}
\alias{coxNMF-package}
\title{coxNMF: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
