% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initializes the coxNMF model}
\usage{
init(
  X,
  M,
  y,
  delta,
  k,
  alpha,
  lambda = 0,
  eta = 0,
  gamma = 0,
  WtX = TRUE,
  BFGS = TRUE,
  penalty = "lasso",
  verbose = FALSE,
  tol = 1e-06,
  ninit = 10,
  imaxit = 15
)
}
\arguments{
\item{X}{matrix of gene expression values of dimension p genes by n samples}

\item{y}{a vector of n survival times}

\item{delta}{a vector of n event indicators 0=censored, 1=died}

\item{k}{the rank of the NMF decomposition}

\item{alpha}{the balancing parameter for NMF and survival contribution to loss}

\item{lambda}{parameter for penalty on the coefficients in survival model}

\item{eta}{parameter for balance of l1 and l2 penalty on the coefficients, eta=1 is all l1}

\item{gamma}{parameter for l1 penalty on W}

\item{WtX}{boolean indicator if TRUE, rows of WtX are the covariates fed to the survival model, if FALSE, rows of H are the covariates}

\item{BFGS}{whether to run BFGS for updating W}

\item{penalty}{the type of penalty on the survival coefficients}

\item{verbose}{if TRUE prints loss at each iteration}

\item{tol}{convergence criteria}

\item{ninit}{number of initializations}

\item{imaxit}{max number of iterations for each initialization}
}
\description{
Outputs the model with the smallest loss value across many random initializations
}
