% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxNMF.R
\name{run_coxNMF}
\alias{run_coxNMF}
\title{Runs the coxNMF model}
\usage{
run_coxNMF(
  X,
  y,
  delta,
  k,
  alpha,
  lambda,
  eta,
  gamma = 0,
  H0 = NULL,
  W0 = NULL,
  beta0 = NULL,
  M = NULL,
  WtX = TRUE,
  verbose = FALSE,
  tol = 1e-06,
  maxit = 500,
  penalty = "lasso",
  BFGS = TRUE,
  ninit = 10,
  imaxit = 15,
  ...
)
}
\arguments{
\item{X}{matrix of gene expression values of dimension p genes by n samples}

\item{y}{a vector of n survival times}

\item{delta}{a vector of n event indicators 0=censored, 1=died}

\item{k}{the rank of the NMF decomposition}

\item{alpha}{the balancing parameter for NMF and survival contribution to loss}

\item{lambda}{parameter for penalty on the coefficients in survival model}

\item{eta}{parameter for balance of l1 and l2 penalty on the coefficients, eta=1 is all l1}

\item{gamma}{parameter for l1 penalty on W}

\item{H0}{matrix of initial values for H of dimension k x n}

\item{W0}{matrix of initial values for W of dimension p x k}

\item{beta0}{vector of k initial values for beta, the survival model coefficients}

\item{M}{matrix of 0s and 1s representing the masking matrix of dimension p x n}

\item{WtX}{boolean indicator if TRUE, rows of WtX are the covariates fed to the survival model, if FALSE, rows of H are the covariates}

\item{verbose}{if TRUE prints loss at each iteration}

\item{tol}{convergence criteria}

\item{maxit}{maximum number of iterations}

\item{penalty}{the type of penalty on the survival coefficients}

\item{BFGS}{whether to run BFGS for updating W}

\item{ninit}{number of initializations, ignored if initializations provided}

\item{imaxit}{max number of iterations for each initialization, ignored if initializations provided}
}
\value{
a list containing the fitted W, H, and beta values, as well as the values of the loss
}
\description{
This function runs the block updating scheme for the coxNMF model for specified parameters. If initialization is not provided, the function first calls the init function.
}
